use nextjs router like a normal person
