--- app/user/[[...slug]]/page.tsx
@@ -1,5 +1,0 @@
-import AuthRouter from '@/components/auth/Router';
-
-export default function UserRouter(props: any) {
-  return <AuthRouter {...props} />;
-}
