+++ app/user/close/page.tsx
@@ -1,0 +1,12 @@
+'use client';
+import { useEffect } from 'react';
+
+export type CloseProps = {};
+
+export function Close() {
+  useEffect(() => {
+    window.close();
+  }, []);
+
+  return null;
+}
