+++ app/user/manage/page.tsx
@@ -17,18 +17,14 @@ export const Profile = ({
   error,
   data,
   router,
-  userDataSWRKey,
   responseMessage,
-  userUpdateEndpoint,
   setResponseMessage,
 }: {
   isLoading: boolean;
   error: any;
   data: any;
   router: any;
-  userDataSWRKey: string;
   responseMessage: string;
-  userUpdateEndpoint: string;
   setResponseMessage: (message: string) => void;
 }) => {
   return (
@@ -83,7 +79,7 @@ export const Profile = ({
             const updateResponse = (
               await axios
                 .put(
-                  `${process.env.NEXT_PUBLIC_AGIXT_SERVER}${userUpdateEndpoint}`,
+                  `${process.env.NEXT_PUBLIC_AGIXT_SERVER}/v1/user`,
                   {
                     ...Object.entries(data).reduce((acc, [key, value]) => {
                       return value ? { ...acc, [key]: value } : acc;
@@ -122,7 +118,7 @@ export const Profile = ({
                 const updateResponse = (
                   await axios
                     .put(
-                      `${process.env.NEXT_PUBLIC_AGIXT_SERVER}${userUpdateEndpoint}`,
+                      `${process.env.NEXT_PUBLIC_AGIXT_SERVER}/v1/user`,
                       {
                         ...data,
                       },
@@ -138,7 +134,7 @@ export const Profile = ({
                 if (updateResponse.detail) {
                   setResponseMessage(updateResponse.detail.toString());
                 }
-                await mutate(userDataSWRKey);
+                await mutate('/user');
                 if (data.missing_requirements && Object.keys(data.missing_requirements).length === 0) {
                   const redirect = getCookie('href') ?? '/';
                   deleteCookie('href');
@@ -154,18 +150,7 @@ export const Profile = ({
   );
 };
 
-export type ManageProps = {
-  userDataSWRKey?: string;
-  userDataEndpoint?: string;
-  userUpdateEndpoint?: string;
-  userPasswordChangeEndpoint?: string;
-};
-
-export default function Manage({
-  userDataSWRKey = '/user',
-  userDataEndpoint = '/v1/user',
-  userUpdateEndpoint = '/v1/user',
-}: ManageProps): ReactNode {
+export default function Manage(): ReactNode {
   const [responseMessage, setResponseMessage] = useState('');
   type User = {
     missing_requirements?: {
@@ -177,9 +162,9 @@ export default function Manage({
     };
   };
   const router = useRouter();
-  const { data, error, isLoading } = useSWR<User, any, string>(userDataSWRKey, async () => {
+  const { data, error, isLoading } = useSWR<User, any, string>('/user', async () => {
     return (
-      await axios.get(`${process.env.NEXT_PUBLIC_AGIXT_SERVER}${userDataEndpoint}`, {
+      await axios.get(`${process.env.NEXT_PUBLIC_AGIXT_SERVER}/v1/user`, {
         headers: {
           'Content-Type': 'application/json',
           Authorization: getCookie('jwt'),
@@ -209,9 +194,7 @@ export default function Manage({
             error,
             data,
             router,
-            userDataSWRKey,
             responseMessage,
-            userUpdateEndpoint,
             setResponseMessage,
           }}
         />
