+++ components/layout/AuthCard.tsx
